Improve Carrier Station Handler

+++ Program/Arduino Main Program/AGV_Arduino_Main_Testing_Final/Motor_Driver.h
@@ -41,7 +41,7 @@ class Motor_Driver{
     void motor_Stop(void);
 
     void set_Dir(motor_dir input_dir);
-    void set_Speed(uint8_t input_speed);
+    void set_Speed(int input_speed);
 };
 
 #endif
